[TASK] Introduce core version based RecordListController override
The `RecordListController` is extended and replaced using XCLASS
technique.

With TYPO3 v12 we now have two things to tackle:

* The controller moved from `ext:record_list` to `ext:backend`
  providing a classmap alias as deprecation layer.
* Method signature for controller method `languageSelector()`
  changed leading to PHP fatal error due to method signature
  incompatibilty.

Even refactored to two controller implementation and using a
corresponding version switch for declaring the used class to
be xlassed, the method signature problem is present. However,
this can mitigated to add a unused second argument for the
TYPO3 v11 RecordListController override class, which is done.

Tasks:

* Rename `\WebVision\WvDeepltranslate\Override\DeeplRecordListController`
  to `\WebVision\WvDeepltranslate\Override\Core11\DeeplRecordListController`.

* Provide the working TYPO3 v12 implementation as
  `\WebVision\WvDeepltranslate\Override\Core12\DeeplRecordListController`.

* Use a unused second untyped argument for the v11 override to avoid
  PHP Fatal Error during compile time.

* Conditionally register xclass for `RecordListController` override in
  `ext_localconf.php`.

Resolves: #252
Releases: main